namespace/registry renamed to namespace/nsregistry, also lint fixes
