+++ common/namespace/nsregistry/fx.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package registry
+package nsregistry
 
 import (
 	"go.uber.org/fx"
