+++ common/namespace/nsregistry/registry_test.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package registry_test
+package nsregistry_test
 
 import (
 	"sync"
@@ -42,7 +42,7 @@ import (
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/namespace"
-	"go.temporal.io/server/common/namespace/registry"
+	"go.temporal.io/server/common/namespace/nsregistry"
 	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/primitives/timestamp"
 )
@@ -53,7 +53,7 @@ type (
 		*require.Assertions
 
 		controller     *gomock.Controller
-		regPersistence *registry.MockPersistence
+		regPersistence *nsregistry.MockPersistence
 		registry       namespace.Registry
 	}
 )
@@ -70,8 +70,8 @@ func (s *registrySuite) TearDownSuite() {}
 func (s *registrySuite) SetupTest() {
 	s.Assertions = require.New(s.T())
 	s.controller = gomock.NewController(s.T())
-	s.regPersistence = registry.NewMockPersistence(s.controller)
-	s.registry = registry.NewRegistry(
+	s.regPersistence = nsregistry.NewMockPersistence(s.controller)
+	s.registry = nsregistry.NewRegistry(
 		s.regPersistence,
 		true,
 		dynamicconfig.GetDurationPropertyFn(time.Second),
@@ -169,7 +169,7 @@ func (s *registrySuite) TestListNamespace() {
 	pageToken := []byte("some random page token")
 
 	s.regPersistence.EXPECT().ListNamespaces(gomock.Any(), &persistence.ListNamespacesRequest{
-		PageSize:       registry.CacheRefreshPageSize,
+		PageSize:       nsregistry.CacheRefreshPageSize,
 		IncludeDeleted: true,
 		NextPageToken:  nil,
 	}).Return(&persistence.ListNamespacesResponse{
@@ -178,7 +178,7 @@ func (s *registrySuite) TestListNamespace() {
 	}, nil)
 
 	s.regPersistence.EXPECT().ListNamespaces(gomock.Any(), &persistence.ListNamespacesRequest{
-		PageSize:       registry.CacheRefreshPageSize,
+		PageSize:       nsregistry.CacheRefreshPageSize,
 		IncludeDeleted: true,
 		NextPageToken:  pageToken,
 	}).Return(&persistence.ListNamespacesResponse{
@@ -264,7 +264,7 @@ func (s *registrySuite) TestRegisterStateChangeCallback_CatchUp() {
 	namespaceNotificationVersion++
 
 	s.regPersistence.EXPECT().ListNamespaces(gomock.Any(), &persistence.ListNamespacesRequest{
-		PageSize:       registry.CacheRefreshPageSize,
+		PageSize:       nsregistry.CacheRefreshPageSize,
 		IncludeDeleted: true,
 		NextPageToken:  nil,
 	}).Return(&persistence.ListNamespacesResponse{
@@ -351,7 +351,7 @@ func (s *registrySuite) TestUpdateCache_TriggerCallBack() {
 	namespaceNotificationVersion++
 
 	s.regPersistence.EXPECT().ListNamespaces(gomock.Any(), &persistence.ListNamespacesRequest{
-		PageSize:       registry.CacheRefreshPageSize,
+		PageSize:       nsregistry.CacheRefreshPageSize,
 		IncludeDeleted: true,
 		NextPageToken:  nil,
 	}).Return(&persistence.ListNamespacesResponse{
@@ -402,7 +402,7 @@ func (s *registrySuite) TestUpdateCache_TriggerCallBack() {
 	namespaceNotificationVersion++
 
 	s.regPersistence.EXPECT().ListNamespaces(gomock.Any(), &persistence.ListNamespacesRequest{
-		PageSize:       registry.CacheRefreshPageSize,
+		PageSize:       nsregistry.CacheRefreshPageSize,
 		IncludeDeleted: true,
 		NextPageToken:  nil,
 	}).Return(&persistence.ListNamespacesResponse{
@@ -470,7 +470,7 @@ func (s *registrySuite) TestGetTriggerListAndUpdateCache_ConcurrentAccess() {
 	entryOld := namespace.FromPersistentState(namespaceRecordOld)
 
 	s.regPersistence.EXPECT().ListNamespaces(gomock.Any(), &persistence.ListNamespacesRequest{
-		PageSize:       registry.CacheRefreshPageSize,
+		PageSize:       nsregistry.CacheRefreshPageSize,
 		IncludeDeleted: true,
 		NextPageToken:  nil,
 	}).Return(&persistence.ListNamespacesResponse{
@@ -567,7 +567,7 @@ func (s *registrySuite) TestRemoveDeletedNamespace() {
 	namespaceNotificationVersion++
 
 	s.regPersistence.EXPECT().ListNamespaces(gomock.Any(), &persistence.ListNamespacesRequest{
-		PageSize:       registry.CacheRefreshPageSize,
+		PageSize:       nsregistry.CacheRefreshPageSize,
 		IncludeDeleted: true,
 		NextPageToken:  nil,
 	}).Return(&persistence.ListNamespacesResponse{
@@ -578,7 +578,7 @@ func (s *registrySuite) TestRemoveDeletedNamespace() {
 	}, nil)
 
 	s.regPersistence.EXPECT().ListNamespaces(gomock.Any(), &persistence.ListNamespacesRequest{
-		PageSize:       registry.CacheRefreshPageSize,
+		PageSize:       nsregistry.CacheRefreshPageSize,
 		IncludeDeleted: true,
 		NextPageToken:  nil,
 	}).Return(&persistence.ListNamespacesResponse{
